Fix sorting of _rev in new doc in code editor view
Background: When creating a new doc the _rev field is
added to the model attributes. If fields are already in
the model then the output of JSON.stringify will have the
_rev field at the end of the string.

This is inconsitent with what happens when a doc is
retrieved via a GET, in which case the _rev will appear
after the _id field.

This patch forces the first two keys of the doc to be
_id and _rev and then copied in the other keys in order.
This means that the view of the document directly after
saving and the view of the document after a GET will be
consistent.

This only happens if _rev does not follow _id in _.keys()